refactor: put top-level binding code in its own dir
