+++ src/chameleon/meson.build
@@ -1,6 +1,10 @@
+use_chameleon = get_option('bind_fortran')
+
 chameleon_sources = files(
   'chameleon',
   'src' / 'generator.rb',
   'src' / 'bind_c.rb',
 )
-chameleon_gen = find_program(chameleon_sources[0])
+
+ruby = find_program('ruby', version: '>=3.0.0', required: use_chameleon)
+chameleon_gen = find_program(chameleon_sources[0], required: use_chameleon)
